+++ lib/stack_appodeal_flutter.dart
@@ -1,5 +1,6 @@
 export 'src/appodeal.dart';
 export 'src/appodeal_ad_type.dart';
+export 'src/appodeal_ad_revenue.dart';
 export 'src/appodeal_banner.dart';
 export 'src/appodeal_banner_size.dart';
 export 'src/appodeal_purchase.dart';